format all configs with 4 spaces
When the config-update workflow runs, it performs json.dump which causes whitespace changes if the config wasn't already in that format. To minimize this, and because it's a nice idea anyway, let's reformat them to the standard format. (From what I can tell, JavaScript JSON.stringify and python json.dump are pretty much aligned, so it doesn't matter which one we use)

New configs generated by issue-to-json will now use 4 spaces instead of 2