make PRs from issue-to-json trigger checks
We added the validate-configs workflow which runs automatically as a check on open PRs.
But we also have the automation process for new deployments, which opens PRs when a New Project Configuration issue is opened. It uses the default token so the PR authored by the Actions bot, which causes the checks to not run. https://github.com/peter-evans/create-pull-request/blob/main/docs/concepts-guidelines.md#triggering-further-workflow-runs

We already have the op-config-updates Github app (i.e. bot). If it authors the PR the checks will run; we just need to provide its credentials as repo Secrets.

While doing this I switched away from using peter-evans/create-pull-request to just using gh CLI commands because: i) it wouldn't work without granting the bot more permissions than are actually needed, ii) gh CLI is what I used in the config-update workflow, iii) I would argue the gh CLI is more idiomatic anyway because it is general knowledge not specific to GH actions

I also made the commit and PR messages more descriptive by passing in the abbreviated name of the deployment and not just the issue number, and I included the keyword "close #<issue number>" which will make the issue automatically close when the PR is merged https://docs.github.com/en/get-started/writing-on-github/working-with-advanced-formatting/using-keywords-in-issues-and-pull-requests#linking-a-pull-request-to-an-issue

Testing done on my fork:
https://github.com/JGreenlee/nrel-openpath-deploy-configs/pull/37
This PR was generated from an New Project Configuration issue, authored by @op-config-updates, checks ran and passed